Go back to spreading props
After much consideration, it gives superior ergonomics and it isn't so
difficult to name our props to avoid collisions with valid HTML
attributes. Also made some changes to proxy up events if the developer
decides to implement one that we are using internally.

Also removed the chai dependency where it still existed.