UIButton: refactor, onPressed now fired regardless of `props.pressed`
Semantically, it didn't make sense for a keydown event to trigger a click
handler, so the behavior and recommendation has been revised such that you
should listen for `onPressed` instead of `onClick` or `onKeyDown` to catch
when the user interacts with the button.

If you really only want to listen for clicks, then providing an `onClick` prop
will still work as expected.