[Breaking Change] Revise UITextualInput prop interface
Previously you could pass in certain input props as naked props,
e.g. `type` but passing them in as `inputProps.type` was also
supported--in retrospect, that's super confusing.

The interface has been revised such that all props going to the
input must be declared inside `inputProps`, e.g.

```
<UITextualInput
    inputProps={
        defaultValue: 'foo'
        placeholder: 'bar',
    } />
```

Misc: `hidePlaceholderOnFocus` is now `true` by default; this follows
the overall effort to have a consistent cross-browser experience, as
the previous default of `false` would lead to differences between the
behavior in IE and the rest of the browsers.