Modify components to not apply custom props to JSX elements
This piggybacks off a change in React 15.2, where they now warn
if garbage is being fed into JSX elements (what will become actual
DOM) that doesn't correlate with a real DOM attribute.

The eventual React plan is to allow everything through to the DOM
without maintaining a map of which attributes are legal for each
DOM node type, so this is an interim measure to catch and fix those
occurrences.