boundless-dialog: add closeOnInsideClick, function forms of close props
All closeOn{condition} props now optionally take a function that
receives the event. The return value will determine if the dialog
is closed.

This gives much more granular control for instances where
the implementor might only want to close the dialog if a particular
type of element is clicked, etc.